Changed headers and data for KvPatchMergeOperation
As per the Orchestrate API, changed the Content-Type header for the JSON
partial merge operation to be application/merge-patch+json as opposed to
the standard application/json.

Also added manual JSON encoding of the body, since Guzzle will not do it
for you unless the Content-Type header is standard.

Fixes #14